[feat & fix] 기존 Int 중 Long 타입이어야 하는 값 Long으로 변경, 장바구니 room 세팅

+++ app/src/main/java/com/nagane/table/data/api/NaganeAPI.kt
@@ -28,13 +28,13 @@ interface NaganeAPI {
     @GET("/to/menuList")
     suspend fun getMenuList(
         @Query("storeCode") storeCode : String,
-        @Query("categoryNo") categoryNo : Int,
+        @Query("categoryNo") categoryNo : Long,
     ): ApiResponse<MenuData>
 
     // 메뉴 상세 정보 받아오기
     @GET("/to/menu")
     suspend fun getMenuDetail(
         @Query("storeCode") storeCode : String,
-        @Query("menuNo") menuNo : Int,
+        @Query("menuNo") menuNo : Long,
     ) : ApiResponse<Map<String, MenuDetail>>
 }
